requirements.txt
bcrypt==3.1.7
bson==0.5.9
cffi==1.14.0
click==7.1.1
docopt==0.6.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
MarkupSafe==1.1.1
nltk==3.5
num2words==0.5.10
numpy==1.18.3
pkg-resources==0.0.0
pycparser==2.20
pymongo==3.10.1
pyspellchecker==0.5.4
python-dateutil==2.8.1
regex==2020.4.4
six==1.14.0
tqdm==4.45.0
Werkzeug==1.0.1
